Update ID types to Long and rename trainingTaskId to id in team entity classes

+++ app/src/main/java/com/sedilant/yambol/data/team/TeamObjectivesEntity.kt
@@ -22,5 +22,5 @@ data class TeamObjectivesEntity(
     val description: String,
     val isFinish: Boolean,
     @ColumnInfo(name = "team_id")
-    val teamId: Int,
+    val teamId: Long,
 )
